refactor(billing): extract AdminBillingModals component (adjust + free-period modals)
